Average Gift - Codechef
Chef has a set S containing N distinct integers.

Chef wants to gift Chefina an array A of any finite length such that the following conditions hold true:

Ai∈S ∀i. In other words, each element of the array A should belong to the set S.
The mean value of all the elements in A is exactly X.
Find whether there exists an array A of finite length satisfying the above conditions.



EXAMPLE:
Sample Input 1 
4
3 2
1 2 3
1 5
3
2 5
4 6
1 5
5
Sample Output 1 
YES
NO
YES
YES

Explanation
Test Case 1: One of the valid arrays is A=[2,2]. Here, 2∈{1,2,3}. Also, mean value of the array is 2+22=2.

Test Case 2: Since all elements of the array can only be equal to 3, the mean value of A cannot be equal to 5.

Test Case 3: One of the valid arrays is A=[4,6]. Here, 4∈{4,6} and 6∈{4,6}. Also, mean value of the array is 4+62=5.

Test Case 4: One of the valid arrays is A=[5]. Here, 5∈{5}. Also, mean value of the array is 51=5.