Degree of Polynomial
Chef has a polynomial in one variable x with N terms. The polynomial looks like A0⋅x0+A1⋅x1+…+AN−2⋅xN−2+AN−1⋅xN−1 where Ai−1 denotes the coefficient of the ith term xi−1 for all (1≤i≤N).

Find the degree of the polynomial.

For each test case, output in a single line, the degree of the polynomial.